+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Currency option is passed to ProductPrice components through props, instead of context
+
 ## [1.0.11] - 2018-12-26
 
 ### Fix